+++ Campaign/resources/views/campaigns/_form.blade.php
@@ -38,6 +38,7 @@ $segmentMap = $segments->flatten()->mapWithKeys(function ($item) {
         "oncePerSession": {!! @json($campaign->once_per_session) !!},
         "active": {!! @json($campaign->active) !!},
         "pageviewRules": {!! @json($campaign->pageview_rules) !!} || [],
+        "pageviewAttributes": {!! @json($campaign->pageview_attributes) !!} || [],
         "countries": {!! @json($selectedCountries) !!},
         "countriesBlacklist": {!! @json($countriesBlacklist ?? 0) !!},
         "allDevices": {!! @json($campaign->getAllDevices()) !!},
@@ -81,7 +82,7 @@ $segmentMap = $segments->flatten()->mapWithKeys(function ($item) {
             }
         ],
         "activationMode": "activate-now",
-    };
+    }
     remplib.campaignForm.bind("#campaign-form", campaign);
 </script>